requirements.txt
# Core dependencies
pandas>=2.0.0
openpyxl>=3.1.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6

# FastAPI and web framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Rate limiting
slowapi>=0.1.9

# Configuration and environment
python-dotenv>=1.0.0

# File handling and utilities
pathlib2>=2.3.7

# Optional database support (for future use)
# sqlalchemy>=2.0.0
# psycopg2-binary>=2.9.0  # PostgreSQL

# Development dependencies (install with pip install -r requirements-dev.txt)
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# httpx>=0.24.0  # For testing FastAPI
# black>=23.0.0  # Code formatting
# isort>=5.12.0  # Import sorting
# flake8>=6.0.0  # Linting